big game code refactor
