+++ backend/app/consumers.py
@@ -1,7 +1,7 @@
 import json
 import asyncio
 from channels.generic.websocket import AsyncWebsocketConsumer
-from .Game import Game
+from .game.Game import Game
 from .models import PongGame
 from channels.db import database_sync_to_async
 from django.utils import timezone
